+++ system/database/drivers/pdo/pdo_driver.php
@@ -42,8 +42,8 @@ class CI_DB_pdo_driver extends CI_DB {
 
 	public $dbdriver = 'pdo';
 
-	// the character used to excape - not necessary for PDO
-	protected $_escape_char = '';
+	// the character used to excape
+	public $_escape_char = '"';
 
 	// clause and character used for LIKE escape sequences
 	protected $_like_escape_str;
@@ -60,10 +60,13 @@ class CI_DB_pdo_driver extends CI_DB {
 	// need to track the pdo driver and options
 	public $pdodriver;
 	public $options = array();
+	protected $driver;
 
 	public function __construct($params)
 	{
 		parent::__construct($params);
+
+		$match = array();
 
 		if (preg_match('/([^;]+):/', $this->dsn, $match) && count($match) == 2)
 		{
@@ -74,7 +77,25 @@ class CI_DB_pdo_driver extends CI_DB {
 		else
 		{
 			// Try to build a complete DSN string from params
-			$this->_connect_string($params);
+			if (strpos($this->hostname, ':'))
+			{
+				// hostname generally would have this prototype
+				// $db['hostname'] = 'pdodriver:host(/Server(/DSN))=hostname(/DSN);';
+				// We need to get the prefix (pdodriver used by PDO).
+				$dsnarray = explode(':', $this->hostname);
+				$this->pdodriver = $dsnarray[0];
+
+				// Extract the hostname from the partial dsn
+				$this->hostname = preg_replace('`(host|server)=`', '', $dsnarray[1]);
+			}
+			else
+			{
+				// Invalid DSN, display an error
+				if ( ! array_key_exists('pdodriver', $params))
+				{
+					show_error('Invalid DB Connection String for PDO');
+				}
+			}
 		}
 
 		// clause and character used for LIKE escape sequences
@@ -99,90 +120,7 @@ class CI_DB_pdo_driver extends CI_DB {
 		$this->_random_keyword = ' RND('.time().')'; // database specific random keyword
 	}
 
-	/**
-	 * Connection String
-	 *
-	 * @param	array
-	 * @return	void
-	 */
-	protected function _connect_string($params)
-	{
-		if (strpos($this->hostname, ':'))
-		{
-			// hostname generally would have this prototype
-			// $db['hostname'] = 'pdodriver:host(/Server(/DSN))=hostname(/DSN);';
-			// We need to get the prefix (pdodriver used by PDO).
-			$dsnarray = explode(':', $this->hostname);
-			$this->pdodriver = $dsnarray[0];
-
-			// End dsn with a semicolon for extra backward compability
-			// if database property was not empty.
-			if ( ! empty($this->database))
-			{
-				$this->dsn .= rtrim($this->hostname, ';').';';
-			}
-		}
-		else
-		{
-			// Invalid DSN, display an error
-			if ( ! array_key_exists('pdodriver', $params))
-			{
-				show_error('Invalid DB Connection String for PDO');
-			}
-
-			// Assuming that the following DSN string format is used:
-			// $dsn = 'pdo://username:password@hostname:port/database?pdodriver=pgsql';
-			$this->dsn = $this->pdodriver.':';
-
-			// Add hostname to the DSN for databases that need it
-			if ( ! empty($this->hostname)
-				&& strpos($this->hostname, ':') === FALSE
-				&& in_array($this->pdodriver, array('informix', 'mysql', 'pgsql', 'sybase', 'mssql', 'dblib', 'cubrid')))
-			{
-			    $this->dsn .= 'host='.$this->hostname.';';
-			}
-
-			// Add a port to the DSN for databases that can use it
-			if ( ! empty($this->port) && in_array($this->pdodriver, array('informix', 'mysql', 'pgsql', 'ibm', 'cubrid')))
-			{
-			    $this->dsn .= 'port='.$this->port.';';
-			}
-		}
-
-		// Add the database name to the DSN, if needed
-	    if (stripos($this->dsn, 'dbname') === FALSE
-	       && in_array($this->pdodriver, array('4D', 'pgsql', 'mysql', 'firebird', 'sybase', 'mssql', 'dblib', 'cubrid')))
-	    {
-	        $this->dsn .= 'dbname='.$this->database.';';
-	    }
-	    elseif (stripos($this->dsn, 'database') === FALSE && in_array($this->pdodriver, array('ibm', 'sqlsrv')))
-	    {
-	    	if (stripos($this->dsn, 'dsn') === FALSE)
-	    	{
-		        $this->dsn .= 'database='.$this->database.';';
-	    	}
-	    }
-	    elseif ($this->pdodriver === 'sqlite' && $this->dsn === 'sqlite:')
-	    {
-	        if ($this->database !== ':memory')
-	        {
-	            if ( ! file_exists($this->database))
-	            {
-	                show_error('Invalid DB Connection string for PDO SQLite');
-	            }
-
-	            $this->dsn .= (strpos($this->database, DIRECTORY_SEPARATOR) !== 0) ? DIRECTORY_SEPARATOR : '';
-	        }
-
-	        $this->dsn .= $this->database;
-	    }
-
-	    // Add charset to the DSN, if needed
-	    if ( ! empty($this->char_set) && in_array($this->pdodriver, array('4D', 'mysql', 'sybase', 'mssql', 'dblib', 'oci')))
-	    {
-	        $this->dsn .= 'charset='.$this->char_set.';';
-	    }
-	}
+	// --------------------------------------------------------------------------
 
 	/**
 	 * Non-persistent database connection
@@ -214,34 +152,43 @@ class CI_DB_pdo_driver extends CI_DB {
 	// --------------------------------------------------------------------
 
 	/**
-	 * PDO connection
+	 * Load the sub driver and connect to the database
 	 *
 	 * @return	object
 	 */
-	public function pdo_connect()
+	protected function pdo_connect()
 	{
-		// Refer : http://php.net/manual/en/ref.pdo-mysql.connection.php
-		if ($this->pdodriver === 'mysql' && ! is_php('5.3.6'))
+		// Load the sub driver for database-specific stuff
+		$driver = strtolower($this->pdodriver);
+
+		// So many libraries for connecting to the same database!
+		// Since SQL Server is a fork of Sybase, this should cover
+		// a lot of bases with one sub-driver
+		if (in_array($driver, array('dblib','mssql','sybase','sqlsrv')))
 		{
-			$this->options[PDO::MYSQL_ATTR_INIT_COMMAND] = "SET NAMES $this->char_set COLLATE '$this->dbcollat'";
+			$driver = 'sqlsrv';
 		}
 
-		// Connecting...
-		try
-		{
-			$db = new PDO($this->dsn, $this->username, $this->password, $this->options);
-		}
-		catch (PDOException $e)
-		{
-			if ($this->db_debug && empty($this->failover))
-			{
-				$this->display_error($e->getMessage(), '', TRUE);
-			}
+		$file = dirname(__FILE__)."/sub_drivers/{$driver}.php";
 
-			return FALSE;
+		// Fallback to odbc if a non-existant or unimplemented
+		// pdo driver is specified
+		if (file_exists($file))
+		{
+			require($file);
+		}
+		else
+		{
+			$driver = 'odbc';
+			require(dirname(__FILE__).'/sub_drivers/odbc.php');
 		}
 
-		return $db;
+		// Instantiate the sub-driver, and
+		// return the connection object
+		$driver_class = "CI_{$driver}_PDO_Driver";
+		$this->driver = new $driver_class($this);
+
+		return $this->conn_id;
 	}
 
 	// --------------------------------------------------------------------
@@ -296,18 +243,9 @@ class CI_DB_pdo_driver extends CI_DB {
 	 */
 	protected function _prep_query($sql)
 	{
-		if ($this->pdodriver === 'pgsql')
-		{
-			// Change the backtick(s) for Postgre
-			$sql = str_replace('`', '"', $sql);
-		}
-		elseif ($this->pdodriver === 'sqlite')
-		{
-			// Change the backtick(s) for SQLite
-			$sql = str_replace('`', '', $sql);
-		}
-
-		return $sql;
+		return (method_exists($this->driver, 'prep_query'))
+			? $this->driver->prep_query($sql)
+			: $sql;
 	}
 
 	// --------------------------------------------------------------------
@@ -450,11 +388,9 @@ class CI_DB_pdo_driver extends CI_DB {
 	 */
 	public function insert_id($name = NULL)
 	{
-		if ($this->pdodriver === 'pgsql' && $name === NULL && $this->version() >= '8.1')
+		if (method_exists($this->driver, 'insert_id'))
 		{
-			$query = $this->query('SELECT LASTVAL() AS ins_id');
-			$query = $query->row();
-			return $query->ins_id;
+			return $this->driver->insert_id($name);
 		}
 
 		return $this->conn_id->lastInsertId($name);
@@ -478,7 +414,9 @@ class CI_DB_pdo_driver extends CI_DB {
 			return 0;
 		}
 
-		$sql = $this->_count_string.$this->protect_identifiers('numrows').' FROM '.$this->protect_identifiers($table, TRUE, NULL, FALSE);
+		$sql = $this->_count_string.$this->protect_identifiers('numrows')
+			.' FROM '.$this->protect_identifiers($table, TRUE, NULL, FALSE);
+
 		$query = $this->query($sql);
 
 		if ($query->num_rows() == 0)
@@ -504,27 +442,7 @@ class CI_DB_pdo_driver extends CI_DB {
 	 */
 	protected function _list_tables($prefix_limit = FALSE)
 	{
-		if ($this->pdodriver == 'pgsql')
-		{
-			// Analog function to show all tables in postgre
-			$sql = "SELECT * FROM information_schema.tables WHERE table_schema = 'public'";
-		}
-		elseif ($this->pdodriver == 'sqlite')
-		{
-			// Analog function to show all tables in sqlite
-			$sql = "SELECT name FROM sqlite_master WHERE type='table' AND name NOT LIKE 'sqlite_%'";
-		}
-		else
-		{
-			$sql = "SHOW TABLES FROM `".$this->database."`";
-		}
-
-		if ($prefix_limit !== FALSE AND $this->dbprefix != '')
-		{
-			return FALSE;
-		}
-
-		return $sql;
+		return $this->driver->list_tables();
 	}
 
 	// --------------------------------------------------------------------
@@ -554,23 +472,7 @@ class CI_DB_pdo_driver extends CI_DB {
 	 */
 	protected function _field_data($table)
 	{
-		if ($this->pdodriver == 'mysql' or $this->pdodriver == 'pgsql')
-		{
-			// Analog function for mysql and postgre
-			return 'SELECT * FROM '.$this->_from_tables($table).' LIMIT 1';
-		}
-		elseif ($this->pdodriver == 'oci')
-		{
-			// Analog function for oci
-			return 'SELECT * FROM '.$this->_from_tables($table).' WHERE ROWNUM <= 1';
-		}
-		elseif ($this->pdodriver == 'sqlite')
-		{
-			// Analog function for sqlite
-			return 'PRAGMA table_info('.$this->_from_tables($table).')';
-		}
-
-		return 'SELECT TOP 1 FROM '.$this->_from_tables($table);
+		return $this->driver->field_data($this->_from_tables($table));
 	}
 
 	// --------------------------------------------------------------------
@@ -662,7 +564,7 @@ class CI_DB_pdo_driver extends CI_DB {
 			$tables = array($tables);
 		}
 
-		return (count($tables) == 1) ? '`'.$tables[0].'`' : '('.implode(', ', $tables).')';
+		return (count($tables) == 1) ? $tables[0] : '('.implode(', ', $tables).')';
 	}
 
 	// --------------------------------------------------------------------
@@ -746,8 +648,9 @@ class CI_DB_pdo_driver extends CI_DB {
 	{
 		$ids   = array();
 		$where = ($where != '' && count($where) >=1) ? implode(" ", $where).' AND ' : '';
+		$final = array();
 
-		foreach ($values as $key => $val)
+		foreach ($values as $val)
 		{
 			$ids[] = $val[$index];
 
@@ -795,6 +698,11 @@ class CI_DB_pdo_driver extends CI_DB {
 	 */
 	protected function _truncate($table)
 	{
+		if (method_exists($this->driver, 'truncate'))
+		{
+			return $this->driver->truncate($table);
+		}
+
 		return $this->_delete($table);
 	}
 
@@ -827,9 +735,7 @@ class CI_DB_pdo_driver extends CI_DB {
 			$conditions .= implode("\n", $like);
 		}
 
-		$limit = ( ! $limit) ? '' : ' LIMIT '.$limit;
-
-		return 'DELETE FROM '.$this->_from_tables($table).$conditions.$limit;
+		return 'DELETE FROM '.$this->_from_tables($table).$conditions;
 	}
 
 	// --------------------------------------------------------------------
@@ -844,21 +750,9 @@ class CI_DB_pdo_driver extends CI_DB {
 	 * @param	int	the offset value
 	 * @return	string
 	 */
-	protected function _limit($sql, $limit, $offset)
+	public function _limit($sql, $limit, $offset)
 	{
-		if ($this->pdodriver == 'cubrid' OR $this->pdodriver == 'sqlite')
-		{
-			$offset = ($offset == 0) ? '' : $offset.', ';
-
-			return $sql.'LIMIT '.$offset.$limit;
-		}
-		else
-		{
-			$sql .= 'LIMIT '.$limit;
-			$sql .= ($offset > 0) ? ' OFFSET '.$offset : '';
-
-			return $sql;
-		}
+		return $this->driver->limit($sql, $limit, $offset);
 	}
 
 	// --------------------------------------------------------------------