+++ system/database/drivers/pdo/pdo_result.php
@@ -130,11 +130,6 @@ class CI_DB_pdo_result extends CI_DB_result {
 	 */
 	public function list_fields()
 	{
-		if ($this->db->db_debug)
-		{
-			return $this->db->display_error('db_unsuported_feature');
-		}
-
 		return FALSE;
 	}